+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Add .unboxed() method to unbox types which can be unboxed into primitive types
 - Add .getParametrizedType() method to get the parametrized type of class
 - Deprecate .toJSONObject() and .toJSONArray() methods in JSONConversions for .encodeObject() usage
+- Add optional metadata to EncodableResult